requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
beanie==1.26.0
motor==3.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
email-validator==2.2.0
boto3==1.35.0
botocore==1.35.0
python-dotenv==1.0.1
psutil==6.1.0
slowapi==0.1.9
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
